signatures working
now add the teacher signature and and fix the csv_files page. also start implementing a backup solution. and start working on the security of the app also fix the pdf color and work on the multline offences and add the period times, alsp work on the directory/create student directories and dont replalce files